requirements.txt
# LLM API
groq==0.4.2

# Embeddings
sentence-transformers==2.3.1

# Vector store
chromadb==0.4.22

# Utilities
python-dotenv==1.0.0
